requirements.txt
# Streamlit & UI
streamlit==1.32.0
streamlit-option-menu==0.3.6
plotly==5.18.0
altair==5.1.2

# FastAPI & Backend
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.1.0

# Database & Authentication
supabase==2.4.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & LLM
openai==1.35.7
tiktoken==0.7.0

# Data Processing & Analysis
numpy==1.26.4
pandas==2.2.0
scipy==1.13.1

# Document Processing
PyMuPDF==1.23.26
pytesseract==0.3.10
Pillow==10.3.0

# Security & Encryption
cryptography==42.0.8

# HTTP
requests==2.33.1
httpx==0.27.0

# Voice (Optional)
elevenlabs==1.2.2

# Logging
structlog==24.2.0
